Applying unnecessary cast patches from #BEANUTILS-55
git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/beanutils/trunk@469743 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/commons/beanutils/ConvertUtilsBean.java
@@ -257,7 +257,7 @@ public class ConvertUtilsBean {
      * @deprecated Register replacement converters for Double.TYPE and
      *  Double.class instead
      */
-    private Double defaultDouble = new Double((double) 0.0);
+    private Double defaultDouble = new Double(0.0);
 
     /**
      * Gets the default value for Double conversions.
@@ -341,7 +341,7 @@ public class ConvertUtilsBean {
      * @deprecated Register replacement converters for Long.TYPE and
      *  Long.class instead
      */
-    private Long defaultLong = new Long((long) 0);
+    private Long defaultLong = new Long(0);
 
     /**
      * Gets the default value for Long conversions.
